requirements.txt
et-xmlfile==1.1.0
numpy==1.21.0
openpyxl==3.0.7
pandas==1.3.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
selenium==3.141.0
six==1.16.0
urllib3==1.26.6
